Give the chunk vector, the ranks and the request set concrete types
Three loose types in hot paths, all of them concrete before the move to the
library:

- The chunk vector was allocated over the abstract Chunk, so the evolve loop
  dispatched on the chunk type at every step.
- The rank and size fields of Comm were declared Integer, which boxes every
  read of a value the handshake touches constantly.
- The required-rank union built a temporary Set of the index box on every
  chunk of every step.

Not measurable on this case: the coupled step time of cavity3D over 20 ranks
varies more between runs than a change of this size.  The allocation per step
is unmoved, which places it elsewhere -- the source receive buffers are
allocated afresh every step.  The Courant numbers are unchanged.

+++ src/scaleTrack/comm.jl
@@ -48,10 +48,10 @@ struct Comm{T<:CommMember}
     communicator::MPI.Comm
     isMaster::Bool
     isHost::Bool
-    rank::Integer
-    jlRank::Integer
-    masterRank::Integer
-    size::Integer
+    rank::Int64
+    jlRank::Int64
+    masterRank::Int64
+    size::Int64
 end
 
 function Comm(member, communicator)
